+++ packages/vm/src/lib.rs
@@ -14,10 +14,11 @@ mod memory;
 mod middleware;
 mod modules;
 mod serde;
+mod size;
 pub mod testing;
 mod traits;
 
-pub use crate::cache::CosmCache;
+pub use crate::cache::{Cache, CacheOptions};
 pub use crate::calls::{
     call_handle, call_handle_raw, call_init, call_init_raw, call_migrate, call_migrate_raw,
     call_query, call_query_raw,
@@ -31,4 +32,5 @@ pub use crate::features::features_from_csv;
 pub use crate::ffi::{FfiError, FfiResult, GasInfo};
 pub use crate::instance::{GasReport, Instance, InstanceOptions};
 pub use crate::serde::{from_slice, to_vec};
+pub use crate::size::Size;
 pub use crate::traits::{Api, Extern, Querier, Storage};