+++ packages/vm/src/modules/mod.rs
@@ -1,3 +1,5 @@
 mod file_system_cache;
+mod in_memory_cache;
 
 pub use file_system_cache::FileSystemCache;
+pub use in_memory_cache::InMemoryCache;